MuratSurenlerU1Capstone/MuratSurenlerU1Capstone/src/main/java/com/example/MuratSurenlerU1Capstone/Model/SalesTax.java
package com.example.MuratSurenlerU1Capstone.Model;

import java.math.BigDecimal;
import java.util.Objects;

public class SalesTax {

    private String state;
    private BigDecimal rate;

    public String getState() {
        return state;
    }

    public void setState(String state) {
        this.state = state;
    }

    public BigDecimal getRate() {
        return rate;
    }

    public void setRate(BigDecimal rate) {
        this.rate = rate;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        SalesTax that = (SalesTax) o;
        return getState() == that.getState() &&
                Objects.equals(getRate(), that.getRate());
    }

    @Override
    public int hashCode() {
        return Objects.hash(getState(), getRate());
    }
}
